scraper/requirements.txt
pandas
lxml
requests
pyyaml
